Refactoring TpcDAQToOffline_module to allow splitter.
This commit refactors TpcDAQToOffline_module so that most of the
module's work is done in a free function, in a new library. This is so
that the same library can be called from the newly-introduce
SplitterInput.

In this commit, SplitterInput is just a stub. The code compiles and
links, but does not yet have any useful behavior.